+++ packages/smarthr-ui/src/components/AppHeader/multilingualization/index.ts
@@ -1,3 +1,2 @@
 export type { Locale } from './types'
-export { translate } from './translate'
 export { localeMap } from './localeMap'